FIX sensors race condition
There is a circular dependency between sensors and workers: Workers have
to find sensors in order to know how to treat a packet, but sensors have
to had an associated worker in order to know where to forward packet.
This way, the first time options are parsed, it could happen to any of
this is NULL.

In future revisions, packets will be forwarded IP-based, and not
sensor-based, so we could revert this commit. At this moment, only an
initial mutex&cond has been placed to stop workers for start before
sensors are defined.

ZZ youtrack related issues:
* MVP-503 Race condition on f2k startup